switch to storing sessions in the context
This patch allows associating a "session key" with a context by calling
`NewSession(ctx)`. When the session is used in a request, the exchange can call
`GetOrCreateSession(ctx)` to make the request in the appropriate session, or create a
new session with the given session key.

Implements https://github.com/ipfs/go-ipfs/issues/7198